app/requirements.txt
# CrewAI and AI-related packages
crewai==1.3.0
crewai-tools==1.2.1
langchain-community==0.4.1
langchain-openai==1.0.1
langchain-core==1.0.2
langchain-text-splitters==0.1.0
openai==1.109.1
instructor==1.12.0

# FastAPI and web server
fastapi==0.120.4
uvicorn[standard]==0.38.0
starlette==0.50.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11

# Authentication and Security
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1

# Pydantic and validation
pydantic[email]==2.12.4
pydantic-settings==2.11.0
email-validator==2.3.0

# Environment and configuration
python-dotenv==1.2.1

# Testing
pytest==8.4.2

# PDF and document processing
pdfplumber==0.11.7
python-docx==1.2.0
pypdf==6.1.3

# Other utilities
requests==2.32.5
httpx==0.28.1
tenacity==9.1.2
python-dateutil==2.9.0.post0